feat(fe): add tabs using parallel routes
