+++ frontend-client/app/(main)/contest/[id]/@tabs/clarification/page.tsx
@@ -1,0 +1,3 @@
+export default function ContestClarification() {
+  return <h1>Contest Clarification</h1>
+}
