+++ frontend-client/app/(main)/contest/[id]/@tabs/page.tsx
@@ -1,0 +1,3 @@
+export default function ContestTab() {
+  return <h1>Contest Tab</h1>
+}
