+++ frontend-client/app/(main)/contest/[id]/@tabs/problem/page.tsx
@@ -1,0 +1,3 @@
+export default function ContestProblem() {
+  return <h1>Contest Problem</h1>
+}
