+++ frontend-client/app/(main)/contest/[id]/@tabs/standings/page.tsx
@@ -1,0 +1,3 @@
+export default function ContestStandings() {
+  return <h1>Contest Standings</h1>
+}
