+++ frontend-client/app/(main)/contest/[id]/@tabs/top/page.tsx
@@ -1,0 +1,3 @@
+export default function ContestTop() {
+  return <h1>Contest Top</h1>
+}
